+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.3
+
+- Add `~` operator as alternative `serve()`.
+
+## 0.0.2
+
+- Add ability to cancel previous request.
+- Add C# style delegate operator `+`(`+=`), `-`(`-=`).
+- Add `<=` operator as alternative `take(box)`.
+
 ## 0.0.1
 
 - Initial version.